# docs/swagger.yaml
basePath: /api
definitions:
...
  controller.RegisterRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
...
  controller.UserPutAvatarRequest:
    properties:
      avatar:
        type: string
    required:
    - avatar
    type: object
  controller.UserPutNicknameRequest:
    properties:
      nickname:
        type: string
    required:
    - nickname
    type: object
  controller.UserPutPasswordRequest:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    required:
    - newPassword
    - oldPassword
    type: object
...
  model.User:
    properties:
      avatar:
        description: 用户头像、创建时；为空字符串;注意只是保存 文件的hash 地址；而不是 URL 地址
        example: ''''''
        type: string
      createAt:
        type: string
      id:
        description: 用户 ID;其类型是一个 uint 类型
        type: integer
      nickname:
        description: 用户中文名称，默认为 ‘’，可后期通过修改用户信息设置
        type: string
      updateAt:
        type: string
      username:
        description: 用户账户；创建时使用；后期不可修改。
        example: whx
        type: string
    required:
    - avatar
    - id
    - username
    type: object
  model.UserResponse:
    properties:
      avatar:
        description: 用户头像、创建时；为空字符串;注意只是保存 文件的hash 地址；而不是 URL 地址
        example: ''''''
        type: string
      createAt:
        type: string
      id:
        description: 用户 ID;其类型是一个 uint 类型
        type: integer
      nickname:
        description: 用户中文名称，默认为 ‘’，可后期通过修改用户信息设置
        type: string
      token:
        description: 认证令牌
        type: string
      updateAt:
        type: string
      username:
        description: 用户账户；创建时使用；后期不可修改。
        example: whx
        type: string
    required:
    - avatar
    - id
    - token
    - username
    type: object
  utils.KResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: |-
    一个 IM 服务器
    支持用户注册、登录、好友管理、群组聊天
    基于 Go + Gin + GORM + Redis
    目前访问服务器任何接口: localhost:8080/api 前端可用先使用代理设置
    使用文件 hash 获取文件地址可用先固定: localhost:8080/api/file/:hash 前端可用看作做;方便后期迁移
  title: go-net
  version: "1.0"
paths:
...
  /users/:id:
    get:
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 根据用户ID获取用户信息
      tags:
      - 用户
  /users/avatar:
    put:
      parameters:
      - description: 修改头像请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.UserPutAvatarRequest'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 用改用户头像;注意是存hash地址；而不是 url 地址;
      tags:
      - 用户
  /users/get:
    get:
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.UserResponse'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 获取用户信息
      tags:
      - 用户
  /users/login:
    post:
      parameters:
      - description: 登录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterRequest'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.UserResponse'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 用户登录
      tags:
      - 用户
  /users/logout:
    get:
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 退出登录
      tags:
      - 用户
  /users/nickname:
    put:
      parameters:
      - description: 修改中文名称请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.UserPutNicknameRequest'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.User'
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 用改用户中文名称;
      tags:
      - 用户
  /users/password:
    put:
      description: 成功后；服务端将清除当前用户token的登录状态
      parameters:
      - description: 修改密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.UserPutPasswordRequest'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 修改用户密码;
      tags:
      - 用户
  /users/register:
    post:
      parameters:
      - description: 注册请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterRequest'
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.User'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 用户注册
      tags:
      - 用户
  /users/users:
    get:
      parameters:
      - default: '""'
        description: 搜索关键字
        in: query
        name: search
        type: string
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.KResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.User'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/utils.KResponse'
      summary: 根据 search 关键字 获取用户列表
      tags:
      - 用户
swagger: "2.0"